Change background color back to white
Removed "View alert" button.  This button takes the user to alert rules configuration which would be confusing to a user.
Removed "Values" area since these are not useful to a user.
Removed "Labels' area.  I think any information that would be useful to a user in this section should be included in the alert text.
Removed the "Silence" button since this takes a user to the configuration page for silencing alerts and doesn't immediately silence the alert which I think is what they would expect..
For the Firing indicator/button (will show "Resolved" in a following email if the alert is configured to notify when the alert is resolved).  I removed the href to go to the alert configuration page.